bitfield issues

+++ src/client.cpp
@@ -188,8 +188,8 @@ Client::Client(const std::string& port1, const std::string& torrent)
 					   
 						if(mb == NULL)
 							exit(1); // not supposed to be null
-						std::cout << "FD Status == 3: Received msg: " << mb->getId() << std::endl;
-					    switch (mb->getId())
+						std::cout << "FD Status == 3: Received msg: " << (uint8_t)mb->getId() << std::endl;
+					    switch ((uint8_t)mb->getId())
 						{
 							case MSG_ID_UNCHOKE: 
 							{		// 1
